refactor: CommentBox, CommentList 간 리팩토링
